old-tests/submission/Infraware/Forms/contents/Forms/input_maxlength.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_maxlength</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='hidden' id='input_text'></p>
	</form>

	<script>
		
		var input_text = document.getElementById("input_text");
		input_text.maxlength = 5;

		if (typeof(input_text.maxlength) == "number") {
			test(function() {
				assert_equals(input_text.maxlength, 5, "maxlength attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("maxlength attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
